+++ examples/components/responsive-navbar-view/responsive-navbar-view.component.jsx
@@ -6,10 +6,10 @@ import { ResponsiveNavbar } from '../../../src/index';
 
 const ResponsiveNavbarView = (props) => {
   const list = [
-    { name: 'Style', href: '/autocomplete' },
-    { name: 'Item 2 longer and longer', href: '/autocomplete' },
-    { name: 'Item 3 even longer and longer', href: '/autocomplete' },
-    { name: 'Item 4', href: '/autocomplete' },
+    { id: 'item1', name: 'Style', href: '/autocomplete' },
+    { id: 'item2', name: 'Item 2 longer and longer', href: '/autocomplete' },
+    { id: 'item3', name: 'Item 3 even longer and longer', href: '/autocomplete' },
+    { id: 'item4', name: 'Item 4', href: '/autocomplete' },
   ];
 
   const activeKey = 2;